+++ src/IR/Instruction.cpp
@@ -17,33 +17,35 @@ void Instruction::Init(Napi::Env env, Napi::Object &exports) {
     exports.Set("Instruction", func);
 }
 
-Napi::Value Instruction::New(Napi::Env env, llvm::Instruction *instruction) {
-    if (llvm::isa<llvm::AllocaInst>(instruction)) {
-        return AllocaInst::New(env, llvm::cast<llvm::AllocaInst>(instruction));
-    } else if (llvm::isa<llvm::BranchInst>(instruction)) {
-        return BranchInst::New(env, llvm::cast<llvm::BranchInst>(instruction));
-    } else if (llvm::isa<llvm::CallInst>(instruction)) {
-        return CallInst::New(env, llvm::cast<llvm::CallInst>(instruction));
-    } else if (llvm::isa<llvm::InvokeInst>(instruction)) {
-        return InvokeInst::New(env, llvm::cast<llvm::InvokeInst>(instruction));
-    } else if (llvm::isa<llvm::LoadInst>(instruction)) {
-        return LoadInst::New(env, llvm::cast<llvm::LoadInst>(instruction));
-    } else if (llvm::isa<llvm::ReturnInst>(instruction)) {
-        return ReturnInst::New(env, llvm::cast<llvm::ReturnInst>(instruction));
-    } else if (llvm::isa<llvm::ResumeInst>(instruction)) {
-        return ResumeInst::New(env, llvm::cast<llvm::ResumeInst>(instruction));
-    } else if (llvm::isa<llvm::SwitchInst>(instruction)) {
-        return SwitchInst::New(env, llvm::cast<llvm::SwitchInst>(instruction));
-    } else if (llvm::isa<llvm::StoreInst>(instruction)) {
-        return StoreInst::New(env, llvm::cast<llvm::StoreInst>(instruction));
-    } else if (llvm::isa<llvm::LandingPadInst>(instruction)) {
-        return LandingPadInst::New(env, llvm::cast<llvm::LandingPadInst>(instruction));
-    } else if (llvm::isa<llvm::UnreachableInst>(instruction)) {
-        return UnreachableInst::New(env, llvm::cast<llvm::UnreachableInst>(instruction));
-    } else if (llvm::isa<llvm::PHINode>(instruction)) {
-        return PHINode::New(env, llvm::cast<llvm::PHINode>(instruction));
+Napi::Value Instruction::New(Napi::Env env, llvm::Instruction *inst) {
+    if (llvm::isa<llvm::AllocaInst>(inst)) {
+        return AllocaInst::New(env, llvm::cast<llvm::AllocaInst>(inst));
+    } else if (llvm::isa<llvm::LoadInst>(inst)) {
+        return LoadInst::New(env, llvm::cast<llvm::LoadInst>(inst));
+    } else if (llvm::isa<llvm::StoreInst>(inst)) {
+        return StoreInst::New(env, llvm::cast<llvm::StoreInst>(inst));
+    } else if (llvm::isa<llvm::GetElementPtrInst>(inst)) {
+        return GetElementPtrInst::New(env, llvm::cast<llvm::GetElementPtrInst>(inst));
+    } else if (llvm::isa<llvm::CallInst>(inst)) {
+        return CallInst::New(env, llvm::cast<llvm::CallInst>(inst));
+    } else if (llvm::isa<llvm::PHINode>(inst)) {
+        return PHINode::New(env, llvm::cast<llvm::PHINode>(inst));
+    } else if (llvm::isa<llvm::LandingPadInst>(inst)) {
+        return LandingPadInst::New(env, llvm::cast<llvm::LandingPadInst>(inst));
+    } else if (llvm::isa<llvm::ReturnInst>(inst)) {
+        return ReturnInst::New(env, llvm::cast<llvm::ReturnInst>(inst));
+    } else if (llvm::isa<llvm::BranchInst>(inst)) {
+        return BranchInst::New(env, llvm::cast<llvm::BranchInst>(inst));
+    } else if (llvm::isa<llvm::SwitchInst>(inst)) {
+        return SwitchInst::New(env, llvm::cast<llvm::SwitchInst>(inst));
+    } else if (llvm::isa<llvm::InvokeInst>(inst)) {
+        return InvokeInst::New(env, llvm::cast<llvm::InvokeInst>(inst));
+    } else if (llvm::isa<llvm::ResumeInst>(inst)) {
+        return ResumeInst::New(env, llvm::cast<llvm::ResumeInst>(inst));
+    } else if (llvm::isa<llvm::UnreachableInst>(inst)) {
+        return UnreachableInst::New(env, llvm::cast<llvm::UnreachableInst>(inst));
     }
-    return constructor.New({Napi::External<llvm::Instruction>::New(env, instruction)});
+    return constructor.New({Napi::External<llvm::Instruction>::New(env, inst)});
 }
 
 bool Instruction::IsClassOf(const Napi::Value &value) {
@@ -63,16 +65,16 @@ Instruction::Instruction(const Napi::CallbackInfo &info) : ObjectWrap(info) {
         throw Napi::TypeError::New(env, ErrMsg::Class::Instruction::constructor);
     }
     auto external = info[0].As<Napi::External<llvm::Instruction>>();
-    instruction = external.Data();
+    inst = external.Data();
 }
 
 llvm::Instruction *Instruction::getLLVMPrimitive() {
-    return instruction;
+    return inst;
 }
 
 Napi::Value Instruction::userBack(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
-    llvm::Instruction *userInst = instruction->user_back();
+    llvm::Instruction *userInst = inst->user_back();
     if (userInst) {
         return Instruction::New(env, userInst);
     }
@@ -81,7 +83,7 @@ Napi::Value Instruction::userBack(const Napi::CallbackInfo &info) {
 
 Napi::Value Instruction::getParent(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
-    llvm::BasicBlock *basicBlock = instruction->getParent();
+    llvm::BasicBlock *basicBlock = inst->getParent();
     if (basicBlock) {
         return BasicBlock::New(env, basicBlock);
     }
@@ -90,7 +92,7 @@ Napi::Value Instruction::getParent(const Napi::CallbackInfo &info) {
 
 Napi::Value Instruction::getModule(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
-    llvm::Module *module = instruction->getModule();
+    llvm::Module *module = inst->getModule();
     if (module) {
         return Module::New(env, module);
     }
@@ -99,7 +101,7 @@ Napi::Value Instruction::getModule(const Napi::CallbackInfo &info) {
 
 Napi::Value Instruction::getFunction(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
-    llvm::Function *function = instruction->getFunction();
+    llvm::Function *function = inst->getFunction();
     if (function) {
         return Function::New(env, function);
     }
@@ -108,14 +110,14 @@ Napi::Value Instruction::getFunction(const Napi::CallbackInfo &info) {
 
 Napi::Value Instruction::getType(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
-    llvm::Type *type = instruction->getType();
+    llvm::Type *type = inst->getType();
     return Type::New(env, type);
 }
 
 void Instruction::setDebugLoc(const Napi::CallbackInfo &info) {
     if (info.Length() == 1 && DebugLoc::IsClassOf(info[0])) {
         llvm::DebugLoc *location = DebugLoc::Extract(info[0]);
-        instruction->setDebugLoc(*location);
+        inst->setDebugLoc(*location);
         return;
     }
     throw Napi::TypeError::New(info.Env(), ErrMsg::Class::Instruction::setDebugLoc);